refactoring things into components
